Move build env related methods to a separate file
