+++ pyodide-build/pyodide_build/tests/test_common.py
@@ -4,135 +4,12 @@ import pytest
 
 from pyodide_build.common import (
     environment_substitute_args,
-    find_matching_wheels,
     find_missing_executables,
-    get_build_environment_vars,
-    get_build_flag,
     get_num_cores,
     make_zip_archive,
     parse_top_level_import_name,
-    platform,
     repack_zip_archive,
-    search_pyodide_root,
 )
-
-
-def test_get_build_flag():
-    assert len(get_build_flag("SIDE_MODULE_LDFLAGS")) > 0
-    assert len(get_build_flag("SIDE_MODULE_CFLAGS")) > 0
-    # n.b. right now CXXFLAGS is empty so don't check length here, just check it returns
-    get_build_flag("SIDE_MODULE_CXXFLAGS")
-
-
-def test_get_build_environment_vars():
-    vars = get_build_environment_vars()
-    assert "SIDE_MODULE_LDFLAGS" in vars
-    assert "SIDE_MODULE_CFLAGS" in vars
-    assert "SIDE_MODULE_CXXFLAGS" in vars
-
-
-def test_wheel_paths():
-    from pathlib import Path
-
-    old_version = "cp38"
-    PYMAJOR = int(get_build_flag("PYMAJOR"))
-    PYMINOR = int(get_build_flag("PYMINOR"))
-    PLATFORM = platform()
-    current_version = f"cp{PYMAJOR}{PYMINOR}"
-    future_version = f"cp{PYMAJOR}{PYMINOR + 1}"
-    strings = []
-
-    for interp in [
-        old_version,
-        current_version,
-        future_version,
-        "py3",
-        "py2",
-        "py2.py3",
-    ]:
-        for abi in [interp, "abi3", "none"]:
-            for arch in [PLATFORM, "linux_x86_64", "any"]:
-                strings.append(f"wrapt-1.13.3-{interp}-{abi}-{arch}.whl")
-
-    paths = [Path(x) for x in strings]
-    assert [x.stem.split("-", 2)[-1] for x in find_matching_wheels(paths)] == [
-        f"{current_version}-{current_version}-{PLATFORM}",
-        f"{current_version}-abi3-{PLATFORM}",
-        f"{current_version}-none-{PLATFORM}",
-        f"{old_version}-abi3-{PLATFORM}",
-        f"py3-none-{PLATFORM}",
-        f"py2.py3-none-{PLATFORM}",
-        "py3-none-any",
-        "py2.py3-none-any",
-        f"{current_version}-none-any",
-    ]
-
-
-def test_search_pyodide_root(tmp_path):
-    pyproject_file = tmp_path / "pyproject.toml"
-    pyproject_file.write_text("[tool.pyodide]")
-    assert search_pyodide_root(tmp_path) == tmp_path
-    assert search_pyodide_root(tmp_path / "subdir") == tmp_path
-    assert search_pyodide_root(tmp_path / "subdir" / "subdir") == tmp_path
-
-    pyproject_file.unlink()
-    with pytest.raises(FileNotFoundError):
-        search_pyodide_root(tmp_path)
-
-
-def test_check_emscripten_version(monkeypatch):
-    from pyodide_build import common
-
-    s = None
-
-    def get_emscripten_version_info():
-        return s
-
-    needed_version = common.emscripten_version()
-    monkeypatch.setattr(
-        common, "get_emscripten_version_info", get_emscripten_version_info
-    )
-    s = """\
-emcc (Emscripten gcc/clang-like replacement + linker emulating GNU ld) 3.1.4 (14cd48e6ead13b02a79f47df1a252abc501a3269)
-clang version 15.0.0 (https://github.com/llvm/llvm-project ce5588fdf478b6af724977c11a405685cebc3d26)
-Target: wasm32-unknown-emscripten
-Thread model: posix
-"""
-    with pytest.raises(
-        RuntimeError,
-        match=f"Incorrect Emscripten version 3.1.4. Need Emscripten version {needed_version}",
-    ):
-        common.check_emscripten_version()
-
-    s = """\
-emcc (Emscripten gcc/clang-like replacement + linker emulating GNU ld) 1.39.20
-clang version 12.0.0 (/b/s/w/ir/cache/git/chromium.googlesource.com-external-github.com-llvm-llvm--project 55fa315b0352b63454206600d6803fafacb42d5e)
-"""
-
-    with pytest.raises(
-        RuntimeError,
-        match=f"Incorrect Emscripten version 1.39.20. Need Emscripten version {needed_version}",
-    ):
-        common.check_emscripten_version()
-
-    s = f"""\
-emcc (Emscripten gcc/clang-like replacement + linker emulating GNU ld) {common.emscripten_version()} (4343cbec72b7db283ea3bda1adc6cb1811ae9a73)
-clang version 15.0.0 (https://github.com/llvm/llvm-project 7effcbda49ba32991b8955821b8fdbd4f8f303e2)
-"""
-    common.check_emscripten_version()
-
-    def get_emscripten_version_info():  # type: ignore[no-redef]
-        raise FileNotFoundError()
-
-    monkeypatch.setattr(
-        common, "get_emscripten_version_info", get_emscripten_version_info
-    )
-
-    with pytest.raises(
-        RuntimeError,
-        match=f"No Emscripten compiler found. Need Emscripten version {needed_version}",
-    ):
-        common.check_emscripten_version()
 
 
 @pytest.mark.parametrize(
@@ -246,39 +123,3 @@ def test_repack_zip_archive(
         assert fh.namelist() == ["a/b.txt", "a/b/c.txt"]
         assert fh.getinfo("a/b.txt").compress_type == expected_compression_type
     assert input_path.stat().st_size == expected_size
-
-
-def test_get_build_environment_vars_host_env(monkeypatch):
-    # host environment variables should have precedence over
-    # variables defined in Makefile.envs
-
-    import os
-
-    get_build_environment_vars.cache_clear()
-    e = get_build_environment_vars()
-    assert e["PYODIDE"] == "1"
-
-    monkeypatch.setenv("HOME", "/home/user")
-    monkeypatch.setenv("PATH", "/usr/bin:/bin")
-    monkeypatch.setenv("PKG_CONFIG_PATH", "/x/y/z:/c/d/e")
-
-    get_build_environment_vars.cache_clear()
-
-    e_host = get_build_environment_vars()
-    assert e_host.get("HOME") == os.environ.get("HOME")
-    assert e_host.get("PATH") == os.environ.get("PATH")
-    assert e_host["PKG_CONFIG_PATH"].endswith("/x/y/z:/c/d/e")
-
-    assert e_host.get("HOME") != e.get("HOME")
-    assert e_host.get("PATH") != e.get("PATH")
-    assert e_host.get("PKG_CONFIG_PATH") != e.get("PKG_CONFIG_PATH")
-
-    get_build_environment_vars.cache_clear()
-
-    monkeypatch.delenv("HOME")
-    monkeypatch.setenv("RANDOM_ENV", "1234")
-
-    get_build_environment_vars.cache_clear()
-    e = get_build_environment_vars()
-    assert "HOME" not in e
-    assert "RANDOM_ENV" not in e
