+++ src/tests/test_pyodide.py
@@ -13,7 +13,7 @@ from pytest_pyodide.server import spawn_web_server
 
 from conftest import DIST_PATH, ROOT_PATH
 from pyodide.code import CodeRunner, eval_code, find_imports, should_quiet  # noqa: E402
-from pyodide_build.common import get_pyodide_root
+from pyodide_build.build_env import get_pyodide_root
 
 
 def _strip_assertions_stderr(messages: Sequence[str]) -> list[str]:
